Use keydown instead of keypress for toggle via keyboard
Apparently, keydown is the actual event in regular clickable elements
that the browser uses to trigger that the element was clicked when the
user interacts with it via keyboard. So if we're going to attach a
keyboard event handler to the toggle element so it always work as if it
was clicked, then it should be keydown. This way for interactive
elements we prevent the double triggering of the event, once in keydown
and then in keypress, therefore cancelling the toggle action.